refactor: move watcher into base builder
