+++ packages/builder/package.json
@@ -12,13 +12,15 @@
     "@nuxt/common": "^2.2.0",
     "@nuxt/webpack": "^2.2.0",
     "@nuxtjs/devalue": "^1.0.1",
+    "chokidar": "^2.0.4",
     "consola": "^1.4.4",
     "fs-extra": "^7.0.0",
     "glob": "^7.1.3",
     "hash-sum": "^1.0.2",
     "lodash": "^4.17.11",
     "pify": "^4.0.1",
-    "serialize-javascript": "^1.5.0"
+    "serialize-javascript": "^1.5.0",
+    "upath": "^1.1.0"
   },
   "publishConfig": {
     "access": "public"
