+++ packages/webpack/package.json
@@ -17,7 +17,6 @@
     "cache-loader": "^1.2.2",
     "caniuse-lite": "^1.0.30000898",
     "chalk": "^2.4.1",
-    "chokidar": "^2.0.4",
     "consola": "^1.4.4",
     "css-loader": "^1.0.0",
     "cssnano": "^4.1.7",
@@ -39,7 +38,6 @@
     "terser-webpack-plugin": "^1.1.0",
     "thread-loader": "^1.2.0",
     "time-fix-plugin": "^2.0.4",
-    "upath": "^1.1.0",
     "url-loader": "^1.1.2",
     "vue-loader": "^15.4.2",
     "webpack": "^4.22.0",
