+++ packages/webpack/src/builder.js
@@ -1,22 +1,12 @@
-import path from 'path'
 import fs from 'fs'
 import pify from 'pify'
-import map from 'lodash/map'
-import debounce from 'lodash/debounce'
-import concat from 'lodash/concat'
-import omit from 'lodash/omit'
-import uniq from 'lodash/uniq'
-import values from 'lodash/values'
-import chokidar from 'chokidar'
 import webpack from 'webpack'
 import MFS from 'memory-fs'
 import webpackDevMiddleware from 'webpack-dev-middleware'
 import webpackHotMiddleware from 'webpack-hot-middleware'
-import upath from 'upath'
 import consola from 'consola'
 
 import {
-  r,
   parallel,
   sequence
 } from '@nuxt/common'
@@ -26,30 +16,23 @@ import { ClientConfig, ServerConfig, PerfLoader } from './config'
 export default class WebpackBuilder {
   constructor(context) {
     this.context = context
-    this.nuxt = context.nuxt
-    this.options = context.nuxt.options
-    this.isStatic = context.isStatic
-    this.plugins = context.plugins
     // Fields that set on build
     this.compilers = []
     this.compilersWatching = []
     this.webpackDevMiddleware = null
     this.webpackHotMiddleware = null
-    this.watchers = {
-      files: null,
-      custom: null,
-      restart: null
-    }
     this.perfLoader = null
 
     // Initialize shared FS and Cache
-    if (this.options.dev) {
+    if (this.context.options.dev) {
       this.mfs = new MFS()
     }
   }
 
   async build() {
-    this.perfLoader = new PerfLoader(this.options)
+    const options = this.context.options
+
+    this.perfLoader = new PerfLoader(options)
 
     const compilersOptions = []
 
@@ -59,12 +42,12 @@ export default class WebpackBuilder {
 
     // Server
     let serverConfig = null
-    if (this.options.build.ssr) {
+    if (options.build.ssr) {
       serverConfig = new ServerConfig(this).config()
       compilersOptions.push(serverConfig)
     }
 
-    for (const p of this.plugins) {
+    for (const p of this.context.plugins) {
       // Client config
       if (!clientConfig.resolve.alias[p.name]) {
         clientConfig.resolve.alias[p.name] = p.src
@@ -82,7 +65,7 @@ export default class WebpackBuilder {
       const compiler = webpack(compilersOption)
 
       // In dev, write files in memory FS
-      if (this.options.dev) {
+      if (options.dev) {
         compiler.outputFileSystem = this.mfs
       }
 
@@ -90,14 +73,14 @@ export default class WebpackBuilder {
     })
 
     // Warmup perfLoader before build
-    if (this.options.build.parallel) {
+    if (options.build.parallel) {
       consola.info('Warming up worker pools')
       this.perfLoader.warmupAll()
       consola.success('Worker pools ready')
     }
 
     // Start Builds
-    const runner = this.options.dev ? parallel : sequence
+    const runner = options.dev ? parallel : sequence
 
     await runner(this.compilers, (compiler) => {
       return this.webpackCompile(compiler)
@@ -107,25 +90,26 @@ export default class WebpackBuilder {
   webpackCompile(compiler) {
     return new Promise(async (resolve, reject) => {
       const name = compiler.options.name
+      const { nuxt, options } = this.context
 
-      await this.nuxt.callHook('build:compile', { name, compiler })
+      await nuxt.callHook('build:compile', { name, compiler })
 
       // Load renderer resources after build
       compiler.hooks.done.tap('load-resources', async (stats) => {
-        await this.nuxt.callHook('build:compiled', {
+        await nuxt.callHook('build:compiled', {
           name,
           compiler,
           stats
         })
 
         // Reload renderer if available
-        this.nuxt.renderer.loadResources(this.mfs || fs)
+        nuxt.renderer.loadResources(this.mfs || fs)
 
         // Resolve on next tick
         process.nextTick(resolve)
       })
 
-      if (this.options.dev) {
+      if (options.dev) {
         // --- Dev Build ---
         // Client Build, watch is started by dev-middleware
         if (compiler.options.name === 'client') {
@@ -133,7 +117,7 @@ export default class WebpackBuilder {
         }
         // Server, build and watch for changes
         this.compilersWatching.push(
-          compiler.watch(this.options.watchers.webpack, (err) => {
+          compiler.watch(options.watchers.webpack, (err) => {
             /* istanbul ignore if */
             if (err) return reject(err)
           })
@@ -145,8 +129,8 @@ export default class WebpackBuilder {
           if (err) {
             return reject(err)
           } else if (stats.hasErrors()) {
-            if (this.options.build.quiet === true) {
-              err = stats.toString(this.options.build.stats)
+            if (options.build.quiet === true) {
+              err = stats.toString(options.build.stats)
             }
             if (!err) {
               // actual errors will be printed by webpack itself
@@ -165,18 +149,20 @@ export default class WebpackBuilder {
   webpackDev(compiler) {
     consola.debug('Adding webpack middleware...')
 
+    const { nuxt: { renderer }, options } = this.context
+
     // Create webpack dev middleware
     this.webpackDevMiddleware = pify(
       webpackDevMiddleware(
         compiler,
         Object.assign(
           {
-            publicPath: this.options.build.publicPath,
+            publicPath: options.build.publicPath,
             stats: false,
             logLevel: 'silent',
-            watchOptions: this.options.watchers.webpack
+            watchOptions: options.watchers.webpack
           },
-          this.options.build.devMiddleware
+          options.build.devMiddleware
         )
       )
     )
@@ -191,87 +177,22 @@ export default class WebpackBuilder {
             log: false,
             heartbeat: 10000
           },
-          this.options.build.hotMiddleware
+          options.build.hotMiddleware
         )
       )
     )
 
     // Inject to renderer instance
-    if (this.nuxt.renderer) {
-      this.nuxt.renderer.webpackDevMiddleware = this.webpackDevMiddleware
-      this.nuxt.renderer.webpackHotMiddleware = this.webpackHotMiddleware
+    if (renderer) {
+      renderer.webpackDevMiddleware = this.webpackDevMiddleware
+      renderer.webpackHotMiddleware = this.webpackHotMiddleware
     }
-
-    // Start watching client files
-    this.watchClient()
-  }
-
-  watchClient() {
-    const src = this.options.srcDir
-    let patterns = [
-      r(src, this.options.dir.layouts),
-      r(src, this.options.dir.store),
-      r(src, this.options.dir.middleware),
-      r(src, `${this.options.dir.layouts}/*.{vue,js}`),
-      r(src, `${this.options.dir.layouts}/**/*.{vue,js}`)
-    ]
-    if (this._nuxtPages) {
-      patterns.push(
-        r(src, this.options.dir.pages),
-        r(src, `${this.options.dir.pages}/*.{vue,js}`),
-        r(src, `${this.options.dir.pages}/**/*.{vue,js}`)
-      )
-    }
-    patterns = map(patterns, upath.normalizeSafe)
-
-    const options = this.options.watchers.chokidar
-    /* istanbul ignore next */
-    const refreshFiles = debounce(() => this.generateRoutesAndFiles(), 200)
-
-    // Watch for src Files
-    this.watchers.files = chokidar
-      .watch(patterns, options)
-      .on('add', refreshFiles)
-      .on('unlink', refreshFiles)
-
-    // Watch for custom provided files
-    let customPatterns = concat(
-      this.options.build.watch,
-      ...values(omit(this.options.build.styleResources, ['options']))
-    )
-    customPatterns = map(uniq(customPatterns), upath.normalizeSafe)
-    this.watchers.custom = chokidar
-      .watch(customPatterns, options)
-      .on('change', refreshFiles)
-  }
-
-  watchServer() {
-    const nuxtRestartWatch = concat(
-      this.options.serverMiddleware
-        .filter(i => typeof i === 'string')
-        .map(this.nuxt.resolver.resolveAlias),
-      this.options.watch.map(this.nuxt.resolver.resolveAlias),
-      path.join(this.options.rootDir, 'nuxt.config.js')
-    )
-
-    this.watchers.restart = chokidar
-      .watch(nuxtRestartWatch, this.options.watchers.chokidar)
-      .on('change', (_path) => {
-        this.watchers.restart.close()
-        const { name, ext } = path.parse(_path)
-        this.nuxt.callHook('watch:fileChanged', this, `${name}${ext}`)
-      })
   }
 
   async unwatch() {
-    for (const watcher in this.watchers) {
-      if (this.watchers[watcher]) {
-        this.watchers[watcher].close()
-      }
+    for (const watching of this.compilersWatching) {
+      watching.close()
     }
-
-    this.compilersWatching.forEach(watching => watching.close())
-
     // Stop webpack middleware
     if (this.webpackDevMiddleware) {
       await this.webpackDevMiddleware.close()
@@ -279,6 +200,6 @@ export default class WebpackBuilder {
   }
 
   forGenerate() {
-    this.isStatic = true
+    this.context.isStatic = true
   }
 }
